REFACTOR: remove legacy auth and splash screens
- Deleted old login, register, and splash screen implementations
- Preparing for new feature-based architecture integration
- Cleaned up unused navigation and UI code
